# Use hex.EncodeToString instead of fmt.Sprintf("%x")