--- a/enterprise/server/remote_execution/filecache/filecache.go
+++ b/enterprise/server/remote_execution/filecache/filecache.go
@@ -2,6 +2,7 @@ package filecache
 
 import (
 	"context"
+	"encoding/hex"
 	"errors"
 	"flag"
 	"fmt"
@@ -495,7 +496,7 @@ func (v *verifiedWriter) Commit() error {
 	if v.file == nil {
 		return status.FailedPreconditionError("writer is closed")
 	}
-	hashStr := fmt.Sprintf("%x", v.csum.Sum(nil))
+	hashStr := hex.EncodeToString(v.csum.Sum(nil))
 	if v.node.GetDigest().GetHash() != hashStr {
 		return status.DataLossErrorf("data checksum %q does not match expected checksum %q", hashStr, v.node.GetDigest().GetHash())
 	}