EncodeToString is faster and more clear.

Benchmark results for digest.Compute:
```
                               │    base3     │                hex3                 │
                               │    sec/op    │    sec/op     vs base               │
DigestCompute/SHA256/1-24        272.2n ± 12%   194.4n ± 13%  -28.58% (p=0.001 n=7)
DigestCompute/BLAKE3/1-24        1.665µ ± 15%   1.293µ ± 10%  -22.34% (p=0.001 n=7)
DigestCompute/SHA256/10-24       235.3n ± 17%   185.9n ± 10%  -20.99% (p=0.001 n=7)
DigestCompute/BLAKE3/10-24       1.605µ ± 19%   1.300µ ±  8%  -19.00% (p=0.001 n=7)
DigestCompute/SHA256/100-24      259.2n ±  9%   241.6n ± 19%   -6.79% (p=0.026 n=7)
DigestCompute/BLAKE3/100-24      1.749µ ± 43%   1.328µ ±  9%  -24.07% (p=0.001 n=7)
DigestCompute/SHA256/1000-24     608.3n ±  9%   559.1n ±  3%   -8.09% (p=0.001 n=7)
DigestCompute/BLAKE3/1000-24     2.375µ ± 43%   1.985µ ± 32%        ~ (p=0.165 n=7)
DigestCompute/SHA256/10000-24    3.983µ ±  2%   3.969µ ± 15%        ~ (p=0.535 n=7)
DigestCompute/BLAKE3/10000-24    4.924µ ±  7%   4.441µ ± 15%        ~ (p=0.209 n=7)
DigestCompute/SHA256/100000-24   39.46µ ± 15%   37.55µ ±  0%   -4.84% (p=0.001 n=7)
DigestCompute/BLAKE3/100000-24   19.68µ ±  3%   19.46µ ±  3%        ~ (p=0.165 n=7)
geomean                          1.859µ         1.599µ        -14.02%

                               │    base3     │                hex3                 │
                               │     B/op     │     B/op      vs base               │
DigestCompute/SHA256/1-24          360.0 ± 0%     400.0 ± 0%  +11.11% (p=0.001 n=7)
DigestCompute/BLAKE3/1-24        10.90Ki ± 0%   10.92Ki ± 0%   +0.19% (p=0.001 n=7)
DigestCompute/SHA256/10-24         360.0 ± 0%     400.0 ± 0%  +11.11% (p=0.001 n=7)
DigestCompute/BLAKE3/10-24       10.90Ki ± 0%   10.92Ki ± 0%   +0.20% (p=0.001 n=7)
DigestCompute/SHA256/100-24        360.0 ± 0%     400.0 ± 0%  +11.11% (p=0.001 n=7)
DigestCompute/BLAKE3/100-24      10.90Ki ± 0%   10.92Ki ± 0%   +0.19% (p=0.001 n=7)
DigestCompute/SHA256/1000-24       360.0 ± 0%     400.0 ± 0%  +11.11% (p=0.001 n=7)
DigestCompute/BLAKE3/1000-24     10.90Ki ± 0%   10.92Ki ± 0%   +0.19% (p=0.001 n=7)
DigestCompute/SHA256/10000-24      360.0 ± 0%     400.0 ± 0%  +11.11% (p=0.001 n=7)
DigestCompute/BLAKE3/10000-24    10.90Ki ± 0%   10.92Ki ± 0%   +0.21% (p=0.001 n=7)
DigestCompute/SHA256/100000-24     360.0 ± 0%     400.0 ± 0%  +11.11% (p=0.001 n=7)
DigestCompute/BLAKE3/100000-24   10.90Ki ± 0%   10.92Ki ± 0%   +0.22% (p=0.001 n=7)
geomean                          1.958Ki        2.066Ki        +5.51%

                               │   base3    │                hex3                │
                               │ allocs/op  │ allocs/op   vs base                │
DigestCompute/SHA256/1-24        6.000 ± 0%   6.000 ± 0%       ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/1-24        7.000 ± 0%   7.000 ± 0%       ~ (p=1.000 n=7) ¹
DigestCompute/SHA256/10-24       6.000 ± 0%   6.000 ± 0%       ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/10-24       7.000 ± 0%   7.000 ± 0%       ~ (p=1.000 n=7) ¹
DigestCompute/SHA256/100-24      6.000 ± 0%   6.000 ± 0%       ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/100-24      7.000 ± 0%   7.000 ± 0%       ~ (p=1.000 n=7) ¹
DigestCompute/SHA256/1000-24     6.000 ± 0%   6.000 ± 0%       ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/1000-24     7.000 ± 0%   7.000 ± 0%       ~ (p=1.000 n=7) ¹
DigestCompute/SHA256/10000-24    6.000 ± 0%   6.000 ± 0%       ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/10000-24    7.000 ± 0%   7.000 ± 0%       ~ (p=1.000 n=7) ¹
DigestCompute/SHA256/100000-24   6.000 ± 0%   6.000 ± 0%       ~ (p=1.000 n=7) ¹
DigestCompute/BLAKE3/100000-24   7.000 ± 0%   7.000 ± 0%       ~ (p=1.000 n=7) ¹
geomean                          6.481        6.481       +0.00%
¹ all samples are equal
```